KAKEULGAE-64 DDL 추가 작성
