+++ src/main/java/kr/kakaocloud/kakeulgae/controller/BookmarkController.java
@@ -1,11 +1,9 @@
 package kr.kakaocloud.kakeulgae.controller;
 
-import java.util.ArrayList;
-import kr.kakaocloud.kakeulgae.domain.dto.BookmarkDto;
+
 import kr.kakaocloud.kakeulgae.service.BookmarkService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
