+++ src/main/java/kr/kakaocloud/kakeulgae/domain/entity/WorkTypeMemberRelation.java
@@ -1,4 +1,4 @@
-package kr.kakaocloud.kakeulgae.domain.entity.member;
+package kr.kakaocloud.kakeulgae.domain.entity;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
@@ -9,15 +9,15 @@ import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
-import kr.kakaocloud.kakeulgae.domain.entity.Region2nd;
+import kr.kakaocloud.kakeulgae.domain.entity.member.Member;
 import lombok.Getter;
 
 @Getter
 @Entity
 @Table(
-    name = "region_member_relatoin"
+    name = "work_type_member_relation"
 )
-public class RegionMemberRelation {
+public class WorkTypeMemberRelation {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -25,11 +25,11 @@ public class RegionMemberRelation {
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "member_id",
-        foreignKey = @ForeignKey(name = "fk_region_member_relation_member_id"))
+        foreignKey = @ForeignKey(name = "fk_work_type_member_relation_member_id"))
     private Member member;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "region_2nd_id",
-        foreignKey = @ForeignKey(name = "fk_region_member_relation_region_2nd_id"))
-    private Region2nd region2nd;
+    @JoinColumn(name = "work_type_id",
+        foreignKey = @ForeignKey(name = "fk_work_type_member_relation_work_type_id"))
+    private WorkType workType;
 }
